requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0

# LLM and Vector Operations
openai>=1.0.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0

# PDF Processing
pdfplumber>=0.10.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Communication
slack-sdk>=3.23.0

# Configuration and Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

# Development
pytest>=7.4.0
black>=23.0.0
